@@ -3,40 +3,50 @@
   "version": "0.1.5",
   "private": true,
   "dependencies": {
+    "-": "^0.0.1",
+    "@babel/preset-env": "^7.24.4",
+    "@babel/preset-react": "^7.24.1",
+    "@babel/preset-typescript": "^7.24.1",
     "@reach/router": "^1.3.3",
     "@react-spring/web": "^9.7.3",
     "@testing-library/jest-dom": "^5.5.0",
-    "@testing-library/react": "^10.0.2",
-    "@testing-library/user-event": "^10.0.2",
+    "@testing-library/react": "^11.2.7",
+    "@testing-library/user-event": "^12.8.3",
     "@types/classnames": "^2.2.10",
-    "@types/jest": "^25.2.1",
-    "@types/lodash": "^4.14.149",
-    "@types/node": "^13.11.1",
-    "@types/reach__router": "^1.3.4",
-    "@types/react": "^16.9.0",
-    "@types/react-dom": "^16.9.0",
-    "@types/react-helmet": "^5.0.15",
-    "axios": "^0.19.2",
+    "@types/jest": "^26.0.24",
+    "@types/lodash": "^4.14.171",
+    "@types/node": "^16.7.13",
+    "@types/reach__router": "^1.3.15",
+    "@types/react": "^18.0.0",
+    "@types/react-dom": "^18.0.0",
+    "@types/react-helmet": "^6.1.4",
+    "axios": "^0.21.1",
+    "babel-loader": "^9.1.3",
     "classnames": "^2.2.6",
-    "i18next": "^19.3.4",
-    "lodash": "^4.17.19",
-    "notyf": "^3.6.0",
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1",
-    "react-helmet": "^6.0.0",
-    "react-hook-form": "^5.3.1",
-    "react-i18next": "^11.3.4",
-    "react-scripts": "3.4.1",
-    "typescript": "^3.8.3"
+    "flubber": "ericbf/flubber#add-type-definitions",
+    "framer-motion": "^11.1.7",
+    "html-webpack-plugin": "^5.6.0",
+    "i18next": "^21.2.2",
+    "lodash": "^4.17.21",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
+    "react-helmet": "^6.1.0",
+    "react-hook-form": "^7.11.0",
+    "react-i18next": "^11.12.0",
+    "react-scripts": "5.0.1",
+    "save-dev": "^0.0.1-security",
+    "typescript": "^5.4.5",
+    "web-vitals": "^3.5.2",
+    "webpack": "^5.91.0",
+    "webpack-cli": "^5.1.4",
+    "webpack-dev-server": "^5.0.4",
+    "worker-loader": "^3.0.8"
   },
   "scripts": {
     "start": "react-scripts start",
     "build": "react-scripts build",
-    "test:dev": "react-scripts test",
-    "test": "react-scripts test --watchAll=false",
-    "eject": "react-scripts eject",
-    "start:styles": "postcss src/assets/css/app.css -o src/App.css",
-    "prestart": "yarn start:styles"
+    "test": "react-scripts test",
+    "eject": "react-scripts eject"
   },
   "eslintConfig": {
     "extends": "react-app"
@@ -57,6 +67,7 @@
     "@fullhuman/postcss-purgecss": "^2.1.0",
     "autoprefixer": "^9.7.6",
     "axios-mock-adapter": "^1.18.1",
+    "css-loader": "^7.1.1",
     "postcss-cli": "^7.1.0",
     "postcss-import": "^12.0.1",
     "purgecss": "^2.1.0",
